Fix format string signedness warnings
GCC 5.0, with its new -Wformat-signedness, warns about the sign being
different between a type and the format string in printf-format
messages. Fix the sign of the format strings and types so that they
match.

https://bugs.freedesktop.org/show_bug.cgi?id=89070

+++ src/udiskslinuxblock.c
@@ -1040,7 +1040,7 @@ escape_fstab (const gchar *source)
         case '\t':
         case '\n':
         case '\\':
-          g_string_append_printf (s, "\\%03o", source[n]);
+          g_string_append_printf (s, "\\%03o", (guint) source[n]);
           break;
 
         default:
@@ -2573,7 +2573,7 @@ handle_format (UDisksBlock           *block,
         {
           if (invocation != NULL)
             g_dbus_method_invocation_return_error (invocation, UDISKS_ERROR, UDISKS_ERROR_FAILED,
-                                                   "Cannot chown %s to uid=%d and gid=%d: %m", tos_dir, caller_uid, caller_gid);
+                                                   "Cannot chown %s to uid=%u and gid=%u: %m", tos_dir, caller_uid, caller_gid);
           if (umount (tos_dir) != 0)
             {
               udisks_warning ("Error unmounting directory %s: %m", tos_dir);
